services/inspection-evaluator-grpc-app/requirements.txt
grpcio==1.74.0
grpcio-tools==1.74.0
grpcio-health-checking==1.73.1
protobuf
SQLAlchemy==2.0.23
asyncpg==0.29.0
python-dotenv
